requirements.txt
playwright==1.49.1
playwright-stealth==1.0.6
questionary==2.0.1
rich==13.9.4
tqdm==4.66.5
pandas==2.2.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
PyYAML==6.0.2
gspread==6.1.2
google-auth==2.34.0
fastapi==0.133.1
uvicorn==0.41.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
SQLAlchemy==2.0.44
psycopg[binary]==3.2.12
redis==6.4.0
celery==5.5.3
pytrends==4.9.2
email-validator==2.3.0
